+++ DependencyInjection/CmsExtension.php
@@ -66,6 +66,7 @@ class CmsExtension extends Extension implements PrependExtensionInterface
 
         // Load the cms bundle config.
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('services.yaml');
         $loader->load('config.yaml');
         $loader->load('cms.yaml');