+++ Model/Menu.php
@@ -11,6 +11,12 @@ final class Menu extends Aggregate
     /** @var string */
     public $name;
 
+    /** @var string */
+    public $language;
+
+    /** @var int */
+    public $website;
+
     /** @var array */
     public $items;
 }